+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - You can now properly open sessions from the search results.
 - Type groups now expand if needed to highlight the current entry in the tree when a character is opened in a tab.
 - When PCs are related to character entries, there is now a "relationship" field.
+- You can now "undeliver" lore on the campaign lore list that was delivered in a session.  It returns it back to that session as undelivered.
 
 ## v1.5.4 - Fixed migration bug 
 ![](https://img.shields.io/badge/release%20date-November%2015%2C%202025-blue)